requirements.txt
django~=5.0
djangorestframework~=3.14.0
celery~=5.3.0
requests~=2.31.0
ujson~=5.9.0
pyyaml~=6.0.0
django-rest-passwordreset>=1.3.0
python-dotenv==1.0.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
redis==5.2.1
Flask==3.1.0
django-jet-reboot==1.3.10
drf-spectacular==0.28.0
pytest==8.3.5
pytest-cov==6.1.1
coverage==7.8.0
pytest-django==4.11.1
rollbar==1.3.0
django-redis==5.4.0
django-cachalot==2.8.0
django-silk==5.3.2
social-auth-app-django==5.4.3
social-auth-core==4.6.1
